# perform initial sync in parallel