@istio-testing 2020-08-12
@stevenctl: The following test **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
analyze-tests_istio | 7fdf8e56d446c41a16cfe59d2879de3a45bb1fa3 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/26419/analyze-tests_istio/447) | `/test analyze-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->