@stevenctl 2020-08-14
Unless the way I set up this benchmark is flawed, I don't think theres much to gain.

With 100 services, 200 endpoints/200 pods (did not set things up for nodes). 
<pre>
~/g/s/i/i/p/p/s/k/controller ❯❯❯ benchstat sync parallel
name                            old time/op  new time/op   delta
Controller_SyncAll/pod-12       17.2µs ± 2%  102.0µs ± 6%  +493.38%  (p=0.000 n=17+19)
Controller_SyncAll/endpoint-12   274µs ± 2%    165µs ± 7%   -39.80%  (p=0.000 n=18+20)
<strike>Controller_SyncAll/node-12      25.2ns ± 0%  136.0ns ± 1%  +439.33%  (p=0.000 n=18+17)</strike>
Controller_SyncAll/service-12   57.7µs ± 1%   84.5µs ±10%   +46.36%  (p=0.000 n=19+20)
</pre>